// src/app/models/administration/user.model.ts
export class UserModel {
    id: number = 0;
    username: string = '';
    name: string = '';
    email: string = '';
    user_type: string = '';
    warehouse_id: number = 0;
    warehouse: string = '';
    is_active: boolean = false;
    password: string = '';
    password_confirmation: string = '';
}